Move specimen management actions and JSPs
